rename utils/static to utils/constants
