+++ src/components/Home/HomeCarousel/homeCarousel.ts
@@ -4,7 +4,7 @@ import Icon from "vue-awesome/components/Icon.vue"
 import Loader from "@/components/Loader.vue"
 import ProgressiveImage from "@/components/ProgressiveImage.vue"
 import useCarouselImages from "@/models/useCarouselImages"
-import { MAIN_COLOR } from "@/utils/static"
+import { MAIN_COLOR } from "@/utils/constants"
 import { getIdFromUrl as getYoutubeIdFromUrl } from "vue-youtube"
 import useCarousel from "@/hooks/useCarousel"
 import useMediaQueries from "@/hooks/useMediaQueries"
