+++ src/components/Home/HomeVideosCarousel/homeVideosCarousel.ts
@@ -3,7 +3,7 @@ import { Carousel, Slide } from "vue-carousel"
 import Icon from "vue-awesome/components/Icon.vue"
 import useVideos from "@/models/useVideos"
 import Loader from "@/components/Loader.vue"
-import { YOUTUBE_CHANNEL_LABEL, MAIN_COLOR } from "@/utils/static"
+import { YOUTUBE_CHANNEL_LABEL, MAIN_COLOR } from "@/utils/constants"
 import useCarousel from "@/hooks/useCarousel"
 import { getIdFromUrl as getYoutubeIdFromUrl } from "vue-youtube"
 import useMediaQueries from "@/hooks/useMediaQueries"
