+++ src/components/Programs/Editions/edition.ts
@@ -1,7 +1,7 @@
 import { defineComponent, ref } from "@vue/composition-api"
 import CollapseTransition from "@ivanv/vue-collapse-transition"
 import Media from "@/components/Media/Media.vue"
-import { MAIN_COLOR } from "@/utils/static"
+import { MAIN_COLOR } from "@/utils/constants"
 import Icon from "vue-awesome/components/Icon.vue"
 import useMediaQueries from "@/hooks/useMediaQueries"
