+++ src/components/SocialNetworks.vue
@@ -24,7 +24,7 @@
 
 <script lang="ts">
 import { defineComponent } from "@vue/composition-api";
-import { MAIN_COLOR } from "@/utils/static";
+import { MAIN_COLOR } from "@/utils/constants";
 import Icon from "vue-awesome/components/Icon.vue";
 import useSocialNetworks from "@/models/useSocialNetworks";
