+++ src/components/layout/Footer/Footer.vue
@@ -35,7 +35,7 @@ import { defineComponent } from "@vue/composition-api";
 import FooterSponsors from "@/components/layout/Footer/FooterSponsors.vue";
 import SocialNetworks from "@/components/SocialNetworks.vue";
 import FooterSuscribeForm from "@/components/layout/Footer/FooterSuscribeForm.vue";
-import { CONTACT_EMAIL } from "@/utils/static";
+import { CONTACT_EMAIL } from "@/utils/constants";
 import useMediaQueries from "@/hooks/useMediaQueries";
 
 export default defineComponent({
