+++ src/models/useNews.ts
@@ -5,7 +5,7 @@ import { NewsPost } from "@/types"
 import useAsyncData from "@/hooks/useAsyncData"
 import { WPResponseItem } from "@/types/wordpressTypes"
 import Urls from "@/utils/urls"
-import { DATE_FORMAT } from "@/utils/static"
+import { DATE_FORMAT } from "@/utils/constants"
 import dayjs from "dayjs"
 import useWpCategories from "@/models/useWpCategories"
 import newsHelpers from "@/utils/news"
