+++ src/utils/constants.ts
@@ -1,0 +1,26 @@
+export const DOMAIN = "cchv.cl"
+export const DEFAULT_TITLE = "CChV"
+export const CONTACT_EMAIL = "contacto@cchv.cl"
+// A specific sponsor is bigger than others; this is due it's category
+export const BIG_SPONSOR_CATEGORY_LABEL = "Financia"
+export const YOUTUBE_CHANNEL_LABEL = "NOTV"
+export const YOUTUBE_CHANNEL_URL =
+  "https://www.youtube.com/channel/UC3aHIBlFx2pWVt1_tvnibCw/videos"
+export const NEWSLETTER_SUSCRIBE_URL =
+  "https://bienaldeartesmediales.us11.list-manage.com/subscribe/post?u=21121cf28dc5ae821ecbf30c9&amp;id=4459a2b85e"
+// Used as default value to 'order' attribute on custom fields
+// used by getCustomFieldFromPost
+export const DEFAULT_ORDER = 0
+
+// Styles
+export const MOBILE_IMG_HEIGHT = "100px"
+export const MAIN_COLOR = "#344284"
+export const DARKER_MAIN_COLOR = "#1a2142"
+export const MEDIA_MOBILE_HEIGHT = "250px"
+
+// Date format compatible with https://github.com/iamkun/dayjs
+
+export const DATE_LOCALE = "es"
+export const DATE_FORMAT = "DD [de] MMMM, YYYY"
+export const SHORT_DATE_FORMAT = "DD [de] MMM"
+export const CUSTOM_FIELDS_DATE_FORMAT = "DD/MM/YYYY"
