+++ src/utils/date.ts
@@ -1,5 +1,5 @@
 import dayjs from "dayjs"
-import { CUSTOM_FIELDS_DATE_FORMAT } from "./static"
+import { CUSTOM_FIELDS_DATE_FORMAT } from "./constants"
 
 // Returns true if dateString is past
 export const dateIsPast = (
