+++ src/utils/news.ts
@@ -3,7 +3,7 @@ import { NewsKeys } from "@/types/customFieldsKeysTypes"
 import { WpImage, WPResponseItem } from "@/types/wordpressTypes"
 import { NewsPost, Activity } from "@/types"
 import dayjs from "dayjs"
-import { DATE_FORMAT, CUSTOM_FIELDS_DATE_FORMAT } from "./static"
+import { DATE_FORMAT, CUSTOM_FIELDS_DATE_FORMAT } from "./constants"
 import { ActivityType } from "@/types/customFieldsTypes"
 import { dateIsPast } from "./date"
 import useMediaQueries from "@/hooks/useMediaQueries"
